Split unicode_align_test crate into checks and diff modules
The crate fused two missions: gating libnu self-checks (encoder
validity, decoder round-trip, length-prediction agreement, casemap
terminator presence, fold idempotence) and non-gating libnu-vs-ICU
divergence reporting. The lib.rs module tree and the flat tests/
listing gave no hint which was which.

Move the libnu probes into a new `checks` module and the comparison
machinery (Diff, Report, run_corpus, fold_icu, lower_icu, compare,
compare_lower, unicode_block) into a new `diff` module. lib.rs is now
a crate-level overview pointing at the two modules. The libnu_ffi
module stays private.

Rename every integration test with a `check_` or `diff_` prefix and
update its `use` path to import from the matching module. The split
is now visible in three places: every directory listing under
`tests/`, every `use unicode_align_test::…` line, and the cargo
filter `cargo test --test 'check_*'` / `--test 'diff_*'`.

No behaviour change. cargo check --all-targets passes.
