+++ src/redisearch_rs/unicode_align_test/tests/check_casemap_read_terminator.rs
@@ -39,7 +39,7 @@
 
 use std::collections::BTreeMap;
 
-use unicode_align_test::{CaseMapTable, CasemapWalk, walk_casemap_expansion};
+use unicode_align_test::checks::{CaseMapTable, CasemapWalk, walk_casemap_expansion};
 use rayon::prelude::*;
 
 const CAP: usize = 16;
